Change type of vars refering to size to the size_t

+++ kernel/allocator.cpp
@@ -9,11 +9,11 @@ namespace kernel {
 static allocator<heap_allocator, 16> a(_end);
 } // namespace kernel
 
-void *operator new(unsigned int size) {
+void *operator new(size_t size) {
     return kernel::a.allocate(size);
 }
 
-void *operator new(unsigned int, void *address) {
+void *operator new(size_t, void *address) {
     return address;
 }
 
@@ -21,11 +21,11 @@ void operator delete(void *address) noexcept {
     kernel::a.free(address);
 }
 
-void operator delete(void *address, unsigned int) noexcept {
+void operator delete(void *address, size_t) noexcept {
     kernel::a.free(address);
 }
 
-void *operator new[](unsigned int size) {
+void *operator new[](size_t size) {
     return kernel::a.allocate(size);
 }
 
@@ -33,7 +33,7 @@ void operator delete[](void *address) noexcept {
     kernel::a.free(address);
 }
 
-void operator delete[](void *address, unsigned int) noexcept {
+void operator delete[](void *address, size_t) noexcept {
     kernel::a.free(address);
 }
